[recode] Account for spilled values of four bytes
Currently, the `recode` functionality assumes that a `live_value` that
is not an alloca should have a default size of 8 bytes. However, we have
encountered cases, like in the Unifico-compiled `loop` example, where a
spilled value can be of size 4.  Thus, we extend the rewrite functions
`_get_val` and `_put_val`, to account for this case as well.