Update vdso templates for CRIU
Even though we use the qemu images from TransProc, we still get a vdso
error during restore. Thus, we follow the instructions in:

https://github.com/ssrg-vt/TransProc/wiki/Updating-vDSO-Template-in-the-Repository

in order to get the updated templates, and we add these templates to
git. If in the future we try to use TransProc in other qemu images or in
bare metal, we need to keep in mind that we may need to repeat the
generation of these templates.