[migration] Unifico-compiled `loop` example x86->arm (#1)
With this patch and a bug fix in LLVM Unifico (blackgeorge-boom/llvm-unifico#31), 
we are able to migrate a unifico-compiled loop example from x86 and continue 
execution to arm, using TransProc.

* [recode] Account for spilled values of four bytes

Currently, the `recode` functionality assumes that a `live_value` that
is not an alloca should have a default size of 8 bytes. However, we have
encountered cases, like in the Unifico-compiled `loop` example, where a
spilled value can be of size 4.  Thus, we extend the rewrite functions
`_get_val` and `_put_val`, to account for this case as well.

* Update vdso templates for CRIU

Even though we use the qemu images from TransProc, we still get a vdso
error during restore. Thus, we follow the instructions in:

https://github.com/ssrg-vt/TransProc/wiki/Updating-vDSO-Template-in-the-Repository

in order to get the updated templates, and we add these templates to
git. If in the future we try to use TransProc in other qemu images or in
bare metal, we need to keep in mind that we may need to repeat the
generation of these templates.