[recode] Account for spilled values of one/two bytes
In X86, an architecture-specific value can be of size 1 byte. If this is
spilled, it is recorded as a live value of size 1. During `recode` this
needs to be handled, as we did previously for the case of 4 bytes. The
difference is that we recode it as a value of 4 bytes, since AArch64
only handles 4 bytes and above. So, if we encoded it as 1 byte, probably
we would run into issues due to AArch64 reading more than 1 byte from
memory.

Regarding correctness, since both architectures, have the same number of
live values, this expansion should not affect correctness, e.g., by
overwriting other live values, but we need to test this more. Finally,
we also add the case of byte size of 2 for completeness.