Add hyperfine for spawn, recode, and restore
The idea here is that hyperfine will do some warmup runs for us, so that
the last run of either spawn or restore, will be a relatively stable.

Also, recode can be measured as an average of multiple runs.

For the case of restore, it uses `setsid` and `script` to simulate a
terminal environment, which is required for CRIU restore, since
hyperfine runs in non-interactive mode.