[common] Add make targets to reduce measurement noise (#14)
* Add targets to reduce measurement noise

Use at the machine that launches qemu with `reduce-noise-{x86,arm}` and
`{reset-reduce-noise-{x86,arm}` respectively.

* Run the time-critical tasks with nice/taskset

This gives them high priority and runs them only on one cpu, to allow
for more robust measurements.

* Add spawn target to launch applications

* Add hyperfine for spawn, recode, and restore

The idea here is that hyperfine will do some warmup runs for us, so that
the last run of either spawn or restore, will be a relatively stable.

Also, recode can be measured as an average of multiple runs.

For the case of restore, it uses `setsid` and `script` to simulate a
terminal environment, which is required for CRIU restore, since
hyperfine runs in non-interactive mode.