[recode] Implement recode without stack transformation
With stack transformation disabled, what recode will essentially do
in terms of stack and registers will be to:
1. Copy the source activations verbatim to the destination memory,
without performing any transformation on the activations.
2. Copy out the innermost activation's callee-saved registers from
source to destination.

Currently, we assume that migration only happens at function boundaries,
so the only live values that won't be in the stack are the CSRs of the
innermost frame. Other registers, may have been clobbered after
returning from the call, so there is no need to copy them out.

Finally, adds option to disable stack transformation.  Activate
(=disable) with --no-transform, when using `crit`. To use from the
experiments makefile, define the NOTRANSFORM variable.