[recode] Add option to completely skip transformation of core/memory/stack/registers
This allows us to perform ablation studies on recode to estimate what is
the impact of any kind of transformation on the application state, by
only performing OS state transformations. We want to use this as a proxy
for a fair comparison between Unifico and Popcorn, so that we keep the
OS state transformation for both binaries, but measure what would be the
performance improvement if we didn't have to transform anything related
to the application state with Unifico.

Activate with `--measure-transform` when running `crit`, or with
`MEASURE_TRANFORM` when using the top-level experiments Makefile.