[recode] Fix how the FP control register is recoded to x86 (#24)
When recoding from arm to x86, it seems that we have to set `cwd` to a
reasonable default. The definition of `cwd` be found in:
https://elixir.bootlin.com/linux/v5.4.205/source/arch/x86/include/asm/fpu/types.h

Otherwise, when moving back to x86, `c_print_results` gives a SIGFPE
error. It seems that the value `0x037f` (round-to-nearest, 64-bit
precision, all exceptions masked), gives a reasonable default.