[migration] Add makefile multi-trip rules
To be executed from the host of QEMU x86. The rules can:
* start/dump/recode the application on x86 QEMU for a first trip,
* restore/dump/recode the application on arm QEMU for a follow-up trip,
* restore/dump/recode the application on x86 QEMU for a second+ trip,
* restore/dump/recode the application on x86 QEMU for a final trip

This also adds functionality to execute arbitrary commands from a local
x86 to a remote arm QEMU, and also to download images from the remote
arm QEMU.

Finally, adds a `safe-restore` target that can be invoked remotely over
ssh, so that it restores the TTY state properly. This was also needed
because we will invoke restore as a background process (`&`).