Backup pages image before benchmarking recode
Through `hyperfine --setup`, before all the runs, we backup the pages
image and then before every run we restore it through `--prepare`. This
is because in the `NOTRANSFORM` optimization, we reuse the source page
as the destination page, so if we want to measure this multiple times,
we need to restore the original source page image.