Page reuse optimization for Unifico (#28)
The main addition of this patch is a `recode` optimization for the case of unifico.

Basically, when we don't want to transform the pages, through the
`--no-transform` flag to `crit`, or the `NOTRANSFORM=1` flag to the
common makefile, we can reuse the source `pages-1.img` as the
destination page. This avoids copying the page, which saves time during
`crit recode`.

This is a nice and simple trick to actually show the benefit of avoiding
the transformation: Unifico only avoids the stack transformation, which
happens in `transform_stack_and_regs`, and by itself it cannot show much
speedup since the whole `recode` procedure is designed around the
assumption that there will be metadata and preprocessing needed to do
the transformation. So, lots of bookkeeping from `recode` is hard to
avoid even though we don't transform the stack. For this reason, a
simple optimization we do it to reuse the page, so we save on copying,
which in theory might not be possible for Popcorn (due to differences in
the various regions inside the page image, which are not guaranteed not
to cause issues if we blindly copy them, while in unifico we imposed as
much alignment as possible).

* Start simplifying recode in the case of no transformation

For now, it just moves the source pages to the destination to avoid
copying. This reuse saves some time already.

* Restore previous pages-1.img when benchmarking recode

This is helpful in our new recode approach where we reuse the src image,
but we need to back it up first and restore it, so that we can run this
measurement multiple times.

* Backup pages image before benchmarking recode

Through `hyperfine --setup`, before all the runs, we backup the pages
image and then before every run we restore it through `--prepare`. This
is because in the `NOTRANSFORM` optimization, we reuse the source page
as the destination page, so if we want to measure this multiple times,
we need to restore the original source page image.

* Add README for benchmarking